ElementWriter - addVector support
